Rename academy materials to materialsUpload
